Małe poprawki

+++ src/main/resources/flyway/migration/V1_2__init_user_metrics.sql
@@ -4,6 +4,6 @@ CREATE TABLE user_metrics (
   height INT,
   calories INT,
   currentWeight DOUBLE PRECISION,
-  activityLevel INT,
+  activityLevel DOUBLE PRECISION,
   targetAction INT
 );
